only resolve DNS on client create
